requirements.txt
websocket_client==0.57.0
pandas==1.2.0
schema==0.7.2
fastparquet==0.5.0